requirements/base.txt
pandas==0.23.0
ipython==6.4.0
tensorflow==1.8.0
Pillow==5.1.0
-e 'git+https://github.com/pdollar/coco.git#egg=pycocotools&subdirectory=PythonAPI'
Cython==0.28.2
lxml==4.2.1
jupyter==1.0.0
matplotlib==2.2.2
scikit-learn==0.19.1
tqdm==4.23.4